pyproject.toml
[tool.poetry]
name = "spotlighter"
version = "0.1.0"
description = ""
authors = ["Teck Chuan Lim <teckchuan@gmail.com>"]
readme = "README.md"
packages = [{include = "spotlighter", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.57.2"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
pymilvus = "^2.5.2"


# Scripts
hydra-core = "^1.3.2"
requests = "^2.32.3"
[tool.poetry.scripts]
app = "src.main:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
